pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "code-agnostic"
version = "0.1.0"
description = "Centralized hub for LLM coding config: MCP, skills, rules, and agents."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click",
  "rich",
  "jsonschema>=4.0",
  "tomli",
  "pyyaml>=6.0",
  "textual>=0.47"
]

[project.optional-dependencies]
dev = [
  "pre-commit>=4.0",
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "jsonschema>=4.0",
  "ruff>=0.8.0",
  "mypy>=1.11"
]

[project.scripts]
code-agnostic = "code_agnostic.__main__:main"
test = "pytest:main"

[tool.setuptools.packages.find]
include = ["code_agnostic*"]

[tool.setuptools.package-data]
"code_agnostic.apps.opencode" = ["schema.json"]
"code_agnostic.apps.cursor" = ["schema.json"]
"code_agnostic.apps.codex" = ["schema.json"]
